feat: avoid copying in `Band`
`Band` is implemented by using an array internally. Previously, when
expanding or progressing the `Band` the items inside it were shifted by
one. This was done to always keep the array filled from the start (from
index 0). This commit changes the implementation of the `Band`, so that
the elements can also start from anywhere in the array and wrap around
it. This means that the head of array can be at the middle for example,
and the last element in the `Band` is the one before the head in the
array. This avoids unnecessary copying, and works with cheap index
operations.